requirements.txt
appdirs==1.4.4
attrs==22.2.0
black==22.12.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-WTF==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
python-decouple==3.6
python-http-client==3.3.7
regex==2022.10.31
requests==2.28.1
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
urllib3==1.26.13
waitress==2.1.2
Werkzeug==2.2.2
WTForms==3.0.1
